Move color functions to utils
